requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==1.10.14
openai>=1.51.0
typing-extensions>=4.10.0
python-dotenv==1.0.1
requests>=2.31.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10